requirements.txt
fastapi==0.104.1
passlib==1.7.4
psycopg2_binary==2.9.5
